i965: fix clear color bo address relocation
Fixes: 7987d041fda0c9 ("i965/surface_state: Emit the clear color address instead of value.")

+++ src/mesa/drivers/dri/i965/brw_wm_surface_state.c
@@ -220,7 +220,7 @@ brw_emit_surface_state(struct brw_context *brw,
    if (clear_bo != NULL) {
       /* Make sure the offset is aligned with a cacheline. */
       assert((clear_offset & 0x3f) == 0);
-      uint32_t *clear_address =
+      uint64_t *clear_address =
             state + brw->isl_dev.ss.clear_color_state_offset;
       *clear_address = brw_state_reloc(&brw->batch,
                                        *surf_offset +
